user_service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Hamadalishah <hammadalis459@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlmodel = "^0.0.22"
python-multipart = "^0.0.12"
uvicorn = "^0.31.1"
passlib = "^1.7.4"
protobuf = "^5.28.2"
aiokafka = "^0.11.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
grpcio-tools = "^1.67.0"
dapr = "^1.14.0"


[tool.poetry.group.dev.dependencies]
types-protobuf = "^5.28.0.20240924"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

